Make sure VM Maestro recognizes it as a "Project."
For VM Maestro to recognizes the VIRL-BOOK repository as a topology project automatically (during Git cloning, for example), a ".project" file is needed.  Note that the ".project" file can live in the root of the repository or in a sub-folder, depending on whether you want users to see multiple projects in VM Maestro or just one project with multiple sub-folders. In other words, each ".project" file consists of a project.